Program now shows the icons when Game#draw_board is called.
Pieces now have a character to show.

Added and corrected instance properties to the pieces classes.

Created a method inside the class Pieces to determine the icon of
each piece.

Corrected an error in Game#create_board, added the variable board at
the end of the method so it is the value it returns.